change files name(index.vue to sample.vue ,search.vue to index.vue)
